generate cli formula via a script (#22)
Currently the pomerium-cli.rb formula is generated via GoReleaser, but
this hasn't worked correctly since the change to build pomerium-cli
separately on macOS and Windows runners. Now the GoReleaser automation
builds only the Linux build, and so it will automatically update the
formula to be Linux-only.

Instead, let's generate this formula using a script, patterned after the
existing desktop-client-release script and the desktop.yaml workflow.