use apparitor github token for desktop workflow (#23)
Currently the desktop.yaml workflow is failing due to a repository rule
requiring all changes to be made by a PR. We should be able to use the
apparitor github account instead, which has permissions to bypass this
rule.

Also add a friendly name for this workflow, and update the generated
commit message.